pyproject.toml
[project]
name = "llm-radio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Niels van Galen Last", email = "nvangalenlast@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dnslib>=0.9.26",
    "dspy>=2.6.27",
    "fastapi[standard]>=0.116.1",
    "gradio>=5.41.1",
    "mlflow>=3.2.0",
    "requests>=2.32.4",
]

[project.scripts]
llm-radio = "llm_radio:main"
run-servers = "llm_radio.cli:run_servers"
chat = "llm_radio.client:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black[jupyter]>=25.1.0",
    "coloredlogs>=15.0.1",
    "isort>=6.0.1",
    "jupyterlab>=4.4.5",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.1.1",
    "requests-mock>=1.12.1",
    "ruff>=0.12.7",
    "tqdm>=4.67.1",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["W505"]

[tool.ruff.lint.isort]
force-sort-within-sections = true

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/llm_radio"]

[tool.coverage.report]
omit = [
    "src/llm_radio/__init__.py",
]
